Set up RuboCop linter (#11)
We have RuboCop set up in this repo to help avoid code smells and syntax
errors, but it's failing because we introduced it late to the party. To
help set some guidelines around Ruby code in this repo, this commit
will:

- Set the config file correctly so we can configure it in future
- Add the recommended
[rubocop-rspec](https://github.com/rubocop/rubocop-rspec) gem to help
write efficient specs
- Run the
[autocorrect](https://docs.rubocop.org/rubocop/usage/auto_correct.html)
command
- Add a new TODO file for existing violations so that the cop passes

Note that the TODO is supposed to be a temporary measure which gives us
time to fix errors over time. Ideally this list will shrink over time
and not be used as a way to silence RuboCop. Alternatively we can fix
all the errors but I don't want to slow this project down.

In a follow up PR we can add this to the CI checks so can save reviewers
time picking up smaller linting errors.