Set up CI linter (#12)
We have RuboCop installed in this repo, but it doesn't run in CI which
means we can merge code without checks. Adding a simple linter at the
beginning of this project will prevent it accumulating technical debt,
and make code review a lot easier.

This will be a follow up to #11 which makes the linter pass.